+++ src/main/java/uned/webtechnologies/shop/services/ProductService.java
@@ -19,7 +19,7 @@ public class ProductService{
 
         return this.productRepository.findAll();
     }
-    public Product getProduct(int id){
+    public Product getProduct(long id){
         return this.productRepository.getOne(id);
     }
